+++ next.config.mjs
@@ -2,6 +2,7 @@
 // const nextConfig = {};
 
 const nextConfig = {
+  serverExternalPackages: ["koffi"],
   eslint: {
     // Warning: This allows production builds to successfully complete even if
     // your project has ESLint errors.